requirements.txt
requests>=2.28.0
PyYAML>=6.0
qrcode[pil]>=7.0.0
PyGithub>=1.58.0